Check code structure

check_code() compares user code to a solution (i.e. model code) and describes the first way that the user code differs. If the user code exactly matches the solution, check_code() returns a customizable success message.

check_code(correct = "{random_praise()} Correct!",
  incorrect = "{message} {random_encourage()}", ..., solution = NULL,
  user = NULL)

Arguments

correct
A character string to display if the student answer matches the solution code. This character string will be run through glue::glue_data with list(correct = TRUE, message = NULL).

incorrect
A character string to display if the student answer matches the solution code. This character string will be run through glue::glue_data with list(correct = FALSE, message = "<STRING>") where message is the error found while comparing the user solution to the known solution.

...
ignored

solution
(Optional) solution code surrounded by quote(), rlang::quo(), or provided as a character string.

user
(Optional) student code to check against the solution surrounded by quote(), rlang::quo(), or provided as a character string.

Value

a grader_result structure from result. If the student answer differs from the solution code, the message will describe the first way that the answer differs, and it will ask the student to try again. If the answer matches the solution code, the message will be the content of the success argument.

Details

check_code() provides a *strict* check in that the user code must exactly match the solution. It is not enough for the user code to be equivalent to the solution code (e.g. to return the same result as the solution).

You can provide solution code for check_code() to use in two ways:

1. Pass code as a character string or a quoted expression to the solution argument of check_code()

2. Make a "-solution" code chunk for the exercise to be checked in a learnr document. There is no need to supply a solution argument for check_code() if you call it from the "-check" chunk of the same exercise. Likewise, there is no need to supply a user argument when you call check_code() from a learnr document (learnr will provide the code that the student submits when it runs check_code().

For best results, name all arguments provided in the solution code.

Examples

# NOT RUN {
grading_demo()
# }